+++ src/Config/Settings.php
@@ -155,8 +155,22 @@ class Settings {
 			function() {
 				$options = get_option( 'kbfos_settings' );
 				?>
-				<input type='text' class='regular-text'  name='kbfos_settings[kbfos_payopts]' value='<?php echo esc_attr( $options['kbfos_payopts'] ); ?>'>
+				<input type='text' class='regular-text'  name='kbfos_settings[kbfos_payopts]' placeholder='Cash, PayPal, etc...' value='<?php echo esc_attr( $options['kbfos_payopts'] ); ?>'>
 				<p class="description">Comma-seperated values accepted.</p>
+				<?php
+			},
+			'pluginPage',
+			'kbfos_pluginPage_section'
+		);
+
+		add_settings_field(
+			'kbfos_place_type',
+			__( 'Place Types', 'text_domain' ),
+			function() {
+				$options = get_option( 'kbfos_settings' );
+				?>
+				<input type='text' class='regular-text'  name='kbfos_settings[kbfos_place_type]' placeholder='Kebab, Pizza, etc...' value='<?php echo esc_attr( $options['kbfos_place_type'] ); ?>'>
+				<p class="description">Used primarily for emojis. Comma-seperated values accepted.</p>
 				<?php
 			},
 			'pluginPage',